+++ mersenne-31/src/lib.rs
@@ -147,7 +147,7 @@ impl PrimeField for Mersenne31 {
         Self::new(n as u32)
     }
 
-    fn from_canonical_usize(n: u64) -> Self {
+    fn from_canonical_usize(n: usize) -> Self {
         Self::new(n as u32)
     }
